remove preview stuff
