+++ source/B2B.Transactions.Infrastructure/B2B.Transactions.Infrastructure.csproj
@@ -17,7 +17,6 @@ limitations under the License.
 
     <PropertyGroup>
         <TargetFramework>net5.0</TargetFramework>
-        <LangVersion>preview</LangVersion>
     </PropertyGroup>
 
     <ItemGroup>
